services/users/requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
debugpy==1.8.20
Django==6.0.1
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
Markdown==3.10
packaging==25.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
swagger-spec-validator==3.0.4
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
